refactor(utils): move 6 cross-cutting helpers from _color to utils (#696)
The element-name formatting and table/shape uniqueness helpers
(_format_element_name/s, _preview_values, _ensure_one_to_one_mapping,
_validate_shape_index_uniqueness, _validate_table_instance_uniqueness) were
parked in _color.py only to break the _validate<->_color cycle. They are
cross-cutting low-level helpers, so utils.py is their natural home — this
removes the "validation helpers living in the color module" wart.

Cycle-safe: they call only each other + third-party (GeoDataFrame, pd,
AnnData, get_table_keys), never _color/_validate, so utils gains no edge to
either. _color imports the two it uses (_ensure_one_to_one_mapping,
_format_element_name) from utils; re-added GeoDataFrame/SpatialElement
imports to utils (dropped earlier when geometry/color moved out).

Verified: no import cycle; ruff + ruff-format + mypy clean; 410 non-visual
tests pass.
